feat: add stopwatch timer and ability to start break
In the TimerView, I added the ability to start a Stopwatch. The timer length and timer value now bind to the parent view, so it can be controlled from higher up.
I also added a button to start a break, which swaps the timer type, and automatically gives users a break equal to 1/5 of their working time.